agents/code-analysis/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "code-analysis-agent"
version = "0.1.0"
description = "Skeleton implementation of the OpenAI Agent SDK containerized runtime"
readme = "AGENTS.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.51.0",
    "typer>=0.12.5",
    "pydantic>=2.9.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "mcp>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "mypy>=1.11.2",
    "pip-tools>=7.4.0",
    "pre-commit>=3.7.1",
    "pytest>=8.3.2",
    "ruff>=0.6.9",
]

[project.scripts]
agent-cli = "agent.__main__:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
pretty = true
